first pass at action consolidation for urlbar
