+++ app/renderer/components/navigation/urlBar.js
@@ -212,19 +212,7 @@ class UrlBar extends React.Component {
   }
 
   onBlur (e) {
-    // We intentionally do not setUrlBarFocused(false) here because
-    // that state is for managing when it should be set if it is active.
-    windowActions.setNavBarUserInput(e.target.value)
-
-    if (!eventElHasAncestorWithClasses(e, ['urlBarSuggestions', 'urlbarForm'])) {
-      this.updateLocationToSuggestion()
-    }
-  }
-
-  updateLocationToSuggestion () {
-    if (this.props.locationValueSuffix.length > 0) {
-      windowActions.setNavBarUserInput(this.props.locationValue + this.props.locationValueSuffix)
-    }
+    windowActions.urlBarOnBlur(getCurrentWindowId(), e.target.value, this.props.locationValue, eventElHasAncestorWithClasses(e, ['urlBarSuggestions', 'urlbarForm']))
   }
 
   get suggestionLocation () {
@@ -304,8 +292,7 @@ class UrlBar extends React.Component {
 
   onFocus (e) {
     this.select()
-    windowActions.setUrlBarFocused(true)
-    windowActions.setUrlBarSelected(true)
+    windowActions.urlBarOnFocus(getCurrentWindowId())
   }
 
   componentWillMount () {
@@ -315,9 +302,8 @@ class UrlBar extends React.Component {
       this.focus()
       this.select()
       windowActions.setRenderUrlBarSuggestions(false)
-      windowActions.setUrlBarFocused(true)
-      windowActions.setUrlBarSelected(true)
       windowActions.setUrlBarActive(true)
+      windowActions.urlBarOnFocus(getCurrentWindowId())
     })
   }
 
@@ -394,14 +380,6 @@ class UrlBar extends React.Component {
     return ['about:', 'file:', 'chrome:', 'view-source:'].includes(protocol)
   }
 
-  get isHTTPPage () {
-    // Whether this page is HTTP or HTTPS. We don't show security indicators
-    // for other protocols like mailto: and about:.
-    const protocol = urlParse(
-      UrlUtil.getLocationIfPDF(this.props.location)).protocol
-    return protocol === 'http:' || protocol === 'https:'
-  }
-
   get shouldRenderUrlBarSuggestions () {
     return this.props.shouldRender === true &&
       this.props.suggestionList && this.props.suggestionList.size > 0
@@ -454,6 +432,12 @@ class UrlBar extends React.Component {
       searchURL = provider.get('search')
     }
 
+    // Whether this page is HTTP or HTTPS. We don't show security indicators
+    // for other protocols like mailto: and about:.
+    const protocol = urlParse(
+      UrlUtil.getLocationIfPDF(location)).protocol
+    const isHTTPPage = protocol === 'http:' || protocol === 'https:'
+
     const props = {}
 
     props.activeTabId = activeTabId
@@ -483,6 +467,7 @@ class UrlBar extends React.Component {
     props.isActive = urlbar.get('active')
     props.isSelected = urlbar.get('selected')
     props.isFocused = urlbar.get('focused')
+    props.isHTTPPage = isHTTPPage
     props.activateSearchEngine = activateSearchEngine
     props.searchSelectEntry = urlbarSearchDetail
     props.autocompleteEnabled = urlbar.getIn(['suggestions', 'autocompleteEnabled'])
@@ -506,7 +491,7 @@ class UrlBar extends React.Component {
           activateSearchEngine={this.props.activateSearchEngine}
           active={this.props.isActive}
           isSecure={this.props.isSecure}
-          isHTTPPage={this.isHTTPPage}
+          isHTTPPage={this.props.isHTTPPage}
           loading={this.props.loading}
           location={this.props.location}
           searchSelectEntry={this.props.searchSelectEntry}
