+++ js/constants/windowConstants.js
@@ -42,8 +42,10 @@ const windowConstants = {
   WINDOW_UNDO_CLOSED_FRAME: _,
   WINDOW_CLEAR_CLOSED_FRAMES: _,
   WINDOW_FRAME_SHORTCUT_CHANGED: _,
+  WINDOW_URL_BAR_ON_BLUR: _,
+  WINDOW_URL_BAR_ON_FOCUS: _,
+  WINDOW_TAB_ON_FOCUS: _,
   WINDOW_SET_URL_BAR_SELECTED: _,
-  WINDOW_SET_URL_BAR_FOCUSED: _,
   WINDOW_SET_SEARCH_DETAIL: _,
   WINDOW_SET_FIND_DETAIL: _,
   WINDOW_SET_BOOKMARK_DETAIL: _, // If set, also indicates that add/edit is shown
